Show translated documentation only when the latest version is translated (#265)
Related to
 - https://github.com/playframework/playframework/issues/8609
 - https://github.com/playframework/playframework.com/issues/252

I changed the behavior of "showing old Japanese documents when accessing the latest document" and changed to "show English documents if there is no corresponding Japanese document for the latest version".

## Background

There used to be a project to translate documents into Japanese, but unfortunately that project has already become inactive. The latest version of the Japanese document is 2.4.

playframework.com gives preference to language over version when accessing the latest documentation. As a result, Japanese users are always shown outdated documents, and I have seen people who are in trouble by referring to information that is already incorrect. I have told many people, "Please do not look at Japanese documents."

And even worse, this is a problem not only in Japanese but also in French etc.